Changed VisitorID to long instead of int since it needs to be 10 digits.

+++ src/VisitorRegistry/Visit.java
@@ -23,7 +23,7 @@ public class Visit {
         ongoing = false;
     }
 
-    public int getVisitorID() {
+    public long getVisitorID() {
         return visitorID;
     }
